+++ src/xtd.forms.native/include/xtd/forms/native/window_styles.h
@@ -217,7 +217,7 @@ constexpr size_t WS_SIZEBOX = WS_THICKFRAME;
 constexpr size_t WS_TILEDWINDOW = WS_OVERLAPPEDWINDOW;
 #endif
 
-inline static xtd::ustring __windows_style_to_string__(int32_t style) {
+inline static xtd::ustring __windows_style_to_string__(size_t style) {
   xtd::ustring result = "WS_OVERLAPPED";
   
   if ((style & WS_MAXIMIZEBOX) == WS_MAXIMIZEBOX) result += " | WS_MAXIMIZEBOX";
